pyproject.toml
[tool.poetry]
name = "python-sql-qa-bot"
version = "0.1.0"
description = ""
authors = [
  "Porimol Chandro <p.chandro@student.uw.edu.pl>"
]
readme = "README.md"
packages = [{include = "app"}]

keywords = [
    "LLMs",
    "GPT",
    "Python",
    "SQL",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
urllib3 = "1.26.14"
cachetools = "5.3.0"
poetry = "1.5.1"
poetry-core = "1.6.1"
click = "^8.1.7"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
qdrant-client = "1.7.0"
fastapi = "0.105.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
uvicorn = "0.25.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.1.2"
jwt = "^1.3.1"
pyjwt = "^2.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
itsdangerous = "^2.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pre-commit = "^3.1.1"
flake8 = "^6.0.0"
mypy = "^1.2.0"
black = "^23.3.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"


[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
